Refactors GatewayPoolsCard to show the transactions modal
