+++ src/content/Dashboards/MyGatewayPools/GatewayPoolDetails.tsx
@@ -8,7 +8,7 @@ import {
   Modal,
   Skeleton,
   Avatar,
-  Fade,
+  Fade
 } from '@mui/material';
 import { gql, useLazyQuery } from '@apollo/client';
 import { useEffect } from 'react';
@@ -25,16 +25,6 @@ import Scrollbar from '@/components/Scrollbar';
 import { stringMiddleTruncate } from '@/utils/stringMiddleTruncate';
 import Link from '@/components/Link';
 import Error from '@/components/Error';
-
-const LabelWrapper = styled(Box)(
-  ({ theme }) => `
-    font-size: ${theme.typography.pxToRem(10)};
-    font-weight: bold;
-    text-transform: uppercase;
-    border-radius: ${theme.general.borderRadiusSm};
-    padding: ${theme.spacing(0.5, 1)};
-  `
-);
 
 const AvatarWrapper = styled(Avatar)(
   ({ theme }) => `
@@ -115,12 +105,13 @@ function GatewayPoolDetails({
             }}
           >
             <Box
+              onTouchEnd={handleClose}
               p={2.5}
               display="flex"
               alignItems="center"
               justifyContent="space-between"
             >
-              <Box onTouchEnd={handleClose}>
+              <Box>
                 <Typography id={title} gutterBottom variant="h4">
                   Transactions
                 </Typography>
@@ -135,19 +126,20 @@ function GatewayPoolDetails({
               </Box>
             </Box>
             <Divider />
-            <Box
-              sx={{
-                flex: 1,
-              }}
-            >
-              {!error && loading ? (
-                <Skeleton variant="rectangular" height="100%" />
-              ) : error ? (
+            <Box sx={{ flex: 1 }}>
+              {error ? (
                 <Error message={error.message} />
               ) : (
                 <Scrollbar>
                   <Timeline>
-                    {data &&
+                    {loading ? (
+                      <>
+                        <TransactionItemLoading />
+                        <TransactionItemLoading />
+                        <TransactionItemLoading />
+                      </>
+                    ) : (
+                      data &&
                       data.myGatewayPoolsDetails.map(
                         ({
                           txHash,
@@ -156,92 +148,26 @@ function GatewayPoolDetails({
                           operation,
                           createdAt
                         }) => (
-                          <TimelineItem
+                          <TransactionItem
                             key={txHash}
-                            sx={{
-                              p: 0
-                            }}
-                          >
-                            <TimelineOppositeContent
-                              sx={{
-                                width: '105px',
-                                flex: 'none'
-                              }}
-                              color="text.secondary"
-                            >
-                              {new Intl.DateTimeFormat('default', {
-                                day: 'numeric',
-                                month: 'short',
-                                year: '2-digit',
-                                hour: '2-digit',
-                                minute: '2-digit',
-                                second: '2-digit'
-                              })
-                                .format(new Date(createdAt))
-                                .replace(',', '')}
-                            </TimelineOppositeContent>
-                            <TimelineSeparator
-                              sx={{
-                                position: 'relative'
-                              }}
-                            >
-                              <TimelineDot
-                                sx={{
-                                  marginTop: 0,
-                                  top: theme.spacing(1.2)
-                                }}
-                                variant="outlined"
-                                color="primary"
-                              />
-                              <TimelineConnector />
-                            </TimelineSeparator>
-                            <TimelineContent
-                              sx={{
-                                pb: 4
-                              }}
-                            >
-                              <LabelWrapper
-                                component="span"
-                                sx={{
-                                  background: `${OPERATION_COLORS[operation].main}`,
-                                  color: `${theme.palette.getContrastText(
-                                    OPERATION_COLORS[operation].dark
-                                  )}`
-                                }}
-                              >
-                                {operation}
-                              </LabelWrapper>
-                              <Typography
-                                sx={{
-                                  pt: 1
-                                }}
-                                variant="h4"
-                                color="text.primary"
-                              >
-                                {`${amount} ${denominator}`}
-                              </Typography>
-                              <Link
-                                href={`https://finder.extraterrestrial.money/mainnet/tx/${txHash}`}
-                                target="_blank"
-                                rel="noopener"
-                                title={txHash}
-                                underline="none"
-                                variant="body2"
-                                overflow="hidden"
-                                display="block"
-                                sx={{
-                                  pt: 0.5,
-                                  '&:hover': {
-                                    textDecoration: 'underline'
-                                  }
-                                }}
-                              >
-                                {stringMiddleTruncate(txHash)}
-                              </Link>
-                            </TimelineContent>
-                          </TimelineItem>
+                            txHash={txHash}
+                            amount={`${amount} ${denominator}`}
+                            operation={operation}
+                            operationColor={OPERATION_COLORS[operation]}
+                            datetime={new Intl.DateTimeFormat('default', {
+                              day: 'numeric',
+                              month: 'short',
+                              year: '2-digit',
+                              hour: '2-digit',
+                              minute: '2-digit',
+                              second: '2-digit'
+                            })
+                              .format(new Date(createdAt))
+                              .replace(',', '')}
+                          />
                         )
-                      )}
+                      )
+                    )}
                   </Timeline>
                 </Scrollbar>
               )}
@@ -254,3 +180,118 @@ function GatewayPoolDetails({
 }
 
 export default GatewayPoolDetails;
+
+const LabelWrapper = styled(Box)(
+  ({ theme }) => `
+    font-size: ${theme.typography.pxToRem(10)};
+    font-weight: bold;
+    text-transform: uppercase;
+    border-radius: ${theme.general.borderRadiusSm};
+    padding: ${theme.spacing(0.5, 1)};
+  `
+);
+
+export const TransactionItem = ({
+  txHash,
+  datetime,
+  operation,
+  operationColor,
+  amount
+}) => {
+  const theme = useTheme();
+
+  return (
+    <TimelineItem
+      key={txHash}
+      sx={{ p: 0 }}
+    >
+      <TimelineOppositeContent
+        sx={{
+          width: '105px',
+          flex: 'none'
+        }}
+        color="text.secondary"
+      >
+        {datetime}
+      </TimelineOppositeContent>
+      <TimelineSeparator sx={{ position: 'relative' }}>
+        <TimelineDot
+          sx={{
+            mt: 0,
+            top: theme.spacing(1.2)
+          }}
+          variant="outlined"
+          color="primary"
+        />
+        <TimelineConnector />
+      </TimelineSeparator>
+      <TimelineContent sx={{ pb: 4 }}>
+        <LabelWrapper
+          component="span"
+          sx={{
+            background: operationColor.main,
+            color: theme.palette.getContrastText(operationColor.dark)
+          }}
+        >
+          {operation}
+        </LabelWrapper>
+        <Typography sx={{ pt: 1 }} variant="h4" color="text.primary">
+          {amount}
+        </Typography>
+        <Link
+          href={`https://finder.extraterrestrial.money/mainnet/tx/${txHash}`}
+          target="_blank"
+          rel="noopener"
+          title={txHash}
+          underline="none"
+          variant="body2"
+          overflow="hidden"
+          display="block"
+          sx={{
+            pt: 0.5,
+            '&:hover': {
+              textDecoration: 'underline'
+            }
+          }}
+        >
+          {stringMiddleTruncate(txHash)}
+        </Link>
+      </TimelineContent>
+    </TimelineItem>
+  );
+};
+
+const TransactionItemLoading = () => {
+  const theme = useTheme();
+  return (
+    <TimelineItem sx={{ p: 0 }}>
+      <TimelineOppositeContent sx={{ width: '105px', flex: 'none' }}>
+        <Skeleton />
+      </TimelineOppositeContent>
+
+      <TimelineSeparator sx={{ position: 'relative' }}>
+        <TimelineDot
+          sx={{ mt: 0, top: theme.spacing(1.2) }}
+          variant="outlined"
+          color="primary"
+        />
+        <TimelineConnector />
+      </TimelineSeparator>
+      <TimelineContent sx={{ pb: 4 }}>
+        <Skeleton>
+          <LabelWrapper>CLAIM</LabelWrapper>
+        </Skeleton>
+        <Skeleton>
+          <Typography sx={{ pt: 1 }} variant="h4" color="text.primary">
+            123124124 UST
+          </Typography>
+        </Skeleton>
+        <Skeleton>
+          <Typography variant="body2" display="block" sx={{ pt: 0.5 }}>
+            141241241241241241
+          </Typography>
+        </Skeleton>
+      </TimelineContent>
+    </TimelineItem>
+  );
+};
