+++ src/content/Dashboards/MyGatewayPools/index.tsx
@@ -1,24 +1,16 @@
 import {
-  Avatar,
   Box,
-  Card,
-  Divider,
   Grid,
   Skeleton,
-  Stack,
-  styled,
-  Typography, useTheme
+  Typography,
 } from '@mui/material';
-import numeral from 'numeral';
-import Decimal from 'decimal.js';
-import StepperProgress from '@/components/StepperProgress';
 import { gql, useLazyQuery } from '@apollo/client';
 import { useWallet, WalletStatus } from '@terra-money/wallet-provider';
 import { useEffect, useState } from 'react';
 import { poolsByIdentifier } from '@/content/DashboardPages/gateway-pools/pools';
-import { amountFormatter } from '@/utils/numberFormatters';
 import Error from '@/components/Error';
 import WalletConnect from '@/components/WalletConnect';
+import { GatewayPoolCard } from '@/content/Dashboards/MyGatewayPools/GatewayPoolCard';
 
 const QUERY = gql`
   query MyGatewayPools($terraWallet: String!) {
@@ -108,7 +100,9 @@ function MyGatewayPools() {
              withdrawAt
            }) => (
             <Grid key={friendlyName} item xs={12} sm={6} md={4} xl={3}>
-              <PoolCard
+              <GatewayPoolCard
+                walletAddress={myWalletAddress}
+                poolContractId={poolContractAddress}
                 title={poolsByIdentifier[poolIdentifier].title}
                 logo={poolsByIdentifier[poolIdentifier].logo}
                 rewardDenominator={rewardDenominator}
@@ -168,150 +162,3 @@ function MyGatewayPools() {
 }
 
 export default MyGatewayPools;
-
-const AvatarWrapper = styled(Avatar)(
-  ({ theme }) => `
-      background: ${theme.colors.alpha.black[100]};
-      width: ${theme.spacing(5.5)};
-      height: ${theme.spacing(5.5)};
-      
-      img {
-        width: 40px;
-        height: 40px;
-        border-radius: 50%;
-      }
-`
-);
-
-export const PoolCard = ({
-  backgroundColor = null,
-  logo,
-  title,
-  rewardDenominator,
-  rewardUAmountDivisor,
-  totalDepositAmount,
-  rewardsUrl,
-  totalClaimedAmount,
-  totalClaimedAmountInUst,
-  claimedAmountToUstMultiplier,
-  startedAt,
-  rewardsClaimableAt,
-  poolEndsAt
-}) => {
-  const theme = useTheme();
-  const [rewards, setRewards] = useState(0);
-  const [rewardsInUst, setRewardsInUst] = useState(0);
-  const [rewardsError, setRewardsError] = useState('');
-  const [loadingRewards, setLoadingRewards] = useState(false);
-
-  useEffect(() => {
-    if (rewardsUrl) {
-      setLoadingRewards(true);
-      fetch(rewardsUrl)
-        .then((res) => res.json())
-        .then(({ result, error }) => {
-          if (result) {
-            const totalRewards = new Decimal(result.amount)
-              .div(rewardUAmountDivisor)
-              .toNumber();
-            setRewards(totalRewards);
-            setRewardsInUst(totalRewards * claimedAmountToUstMultiplier);
-          } else if (error) {
-            setRewardsError('Error fetching data');
-          }
-        })
-        .finally(() => {
-          setLoadingRewards(false);
-        });
-    }
-  }, []);
-
-  return (
-    <Card
-      sx={{
-        backgroundColor: backgroundColor || theme.colors.alpha.white[100],
-        px: 2,
-        py: 3
-      }}
-    >
-      <Box display="flex" alignItems="center">
-        <AvatarWrapper>
-          <img src={logo} alt={title} />
-        </AvatarWrapper>
-        <Typography
-          sx={{
-            ml: 1.5,
-            fontWeight: 'normal'
-          }}
-          variant="h3"
-          component="h3"
-        >
-          {title}
-        </Typography>
-      </Box>
-      <Box sx={{ mt: 1.5 }}>
-        <Stack
-          divider={<Divider orientation="horizontal" flexItem />}
-          spacing={0}
-        >
-          <Box py={1}>
-            <Typography variant="subtitle2">{'Deposits'}</Typography>
-            <Typography variant="h3">
-              {numeral(totalDepositAmount).format('0,000.[00]')} UST
-            </Typography>
-            <Typography variant="subtitle1">
-              {numeral(totalDepositAmount).format('$0,000.[0000]')}
-            </Typography>
-          </Box>
-
-          <Box py={1}>
-            <Typography variant="subtitle2">{'Pending rewards'}</Typography>
-            {!loadingRewards && rewardsError ? (
-              <Typography variant="h3" color="error">
-                {rewardsError}
-              </Typography>
-            ) : (
-              <>
-                <Typography variant="h3">
-                  {loadingRewards ? (
-                    <Skeleton />
-                  ) : (
-                    `${amountFormatter(rewards)} ${rewardDenominator}`
-                  )}
-                </Typography>
-                <Typography
-                  variant="subtitle1"
-                  sx={{ display: 'inline-block' }}
-                >
-                  {loadingRewards ? (
-                    <Skeleton />
-                  ) : (
-                    numeral(rewardsInUst).format('$0,000.[0000]')
-                  )}
-                </Typography>
-              </>
-            )}
-          </Box>
-
-          <Box py={1}>
-            <Typography variant="subtitle2">{'Claimed rewards'}</Typography>
-            <Typography variant="h3">
-              {`${amountFormatter(totalClaimedAmount)} ${rewardDenominator}`}
-            </Typography>
-            <Typography variant="subtitle1">
-              {numeral(totalClaimedAmountInUst).format('$0,000.[0000]')}
-            </Typography>
-          </Box>
-
-          <Box mt={4}>
-            <StepperProgress
-              startedAt={startedAt}
-              rewardsClaimableAt={rewardsClaimableAt}
-              poolEndsAt={poolEndsAt}
-            />
-          </Box>
-        </Stack>
-      </Box>
-    </Card>
-  );
-};
