+++ src/content/Overview/Highlights/data.ts
@@ -1,0 +1,157 @@
+import { GatewayPoolsEnum } from '@/enums/gatewayPools';
+
+export const poolCardsHighlights = [
+  {
+    id: GatewayPoolsEnum['MINE'],
+    rewards: 25080,
+    rewardsInUst: 1.1673,
+    rewardDenominator: 'MINE',
+    totalDepositAmount: 3098.120013,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 8980.998701,
+    totalClaimedAmountInUst: 431.519025585648,
+    claimedAmountToUstMultiplier: 0.048048,
+    startedAt: '2021-07-02T03:00:00.000Z',
+    claimAt: '2022-01-02T03:00:00.000Z',
+    withdrawAt: '2022-07-02T03:00:00.000Z',
+    top: -50,
+    left: -30,
+    animation: '2.1s 2.1s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['LOOP'],
+    rewards: 1940,
+    rewardsInUst: 324,
+    rewardDenominator: 'LOOP',
+    totalDepositAmount: 1198.573402,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 0,
+    totalClaimedAmountInUst: 0.000000,
+    claimedAmountToUstMultiplier: 0.248624,
+    startedAt: '2021-08-13T03:00:00.000Z',
+    claimAt: '2022-05-13T03:00:00.000Z',
+    withdrawAt: '2023-02-13T03:00:00.000Z',
+    top: 515,
+    left: 1152,
+    animation: '2.1s 1.5s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['ORION'],
+    rewards: 25080,
+    rewardsInUst: 1.1673,
+    rewardDenominator: 'ORION',
+    totalDepositAmount: 1258.2213,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 8980.998701,
+    totalClaimedAmountInUst: 431.519025585648,
+    claimedAmountToUstMultiplier: 0.048048,
+    startedAt: '2021-07-02T03:00:00.000Z',
+    claimAt: '2022-01-02T03:00:00.000Z',
+    withdrawAt: '2022-07-02T03:00:00.000Z',
+    top: 280,
+    left: -15,
+    animation: '3.5s 2.4s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['TERRA_WORLD'],
+    rewards: 6440,
+    rewardsInUst: 374.367,
+    rewardDenominator: 'TWD',
+    totalDepositAmount: 1103.979704,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 5200.77821,
+    totalClaimedAmountInUst: 250.000000,
+    claimedAmountToUstMultiplier: 0.063119,
+    startedAt: "2021-08-22T14:00:00.000Z",
+    claimAt: "2022-05-22T14:00:00.000Z",
+    withdrawAt: "2023-02-22T14:00:00.000Z",
+    top: 15,
+    left: 360,
+    animation: '1.5s 1s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['WHITE_WHALE'],
+    rewards: 0,
+    rewardsInUst: 0,
+    rewardDenominator: 'WHALE',
+    totalDepositAmount: 5128.2203,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 12980.924,
+    totalClaimedAmountInUst: 631.619028,
+    claimedAmountToUstMultiplier: 0.048048,
+    startedAt: '2021-07-02T03:00:00.000Z',
+    claimAt: '2022-01-02T03:00:00.000Z',
+    withdrawAt: '2022-07-02T03:00:00.000Z',
+    top: 210,
+    left: 1090,
+    animation: '2.5s 3s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['NEXUS'],
+    rewards: 0,
+    rewardsInUst: 0,
+    rewardDenominator: 'PSI',
+    totalDepositAmount: 2204.449847,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 836.854831,
+    totalClaimedAmountInUst: 27.781906679538,
+    claimedAmountToUstMultiplier: 0.033198,
+    startedAt: '2021-10-05T03:00:00.000Z',
+    claimAt: '2021-10-05T03:00:00.000Z',
+    withdrawAt: '2023-10-05T03:00:00.000Z',
+    top: 480,
+    left: 700,
+    animation: '2.5s 3s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['GLOW'],
+    rewards: 0,
+    rewardsInUst: 0,
+    rewardDenominator: 'GLOW',
+    totalDepositAmount: 501.107147,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 0,
+    totalClaimedAmountInUst: 0.000000,
+    claimedAmountToUstMultiplier: 0.074661,
+    startedAt: '2021-12-19T14:00:00.000Z',
+    claimAt: '2022-09-19T14:00:00.000Z',
+    withdrawAt: '2023-06-19T14:00:00.000Z',
+    top: 295,
+    left: 458,
+    animation: '3s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['SAYVE'],
+    rewards: 0,
+    rewardsInUst: 0,
+    rewardDenominator: 'SAYVE',
+    totalDepositAmount: 751.397377,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 0,
+    totalClaimedAmountInUst: 0.000,
+    claimedAmountToUstMultiplier: 0.001,
+    startedAt: '2021-12-21T14:00:00.000Z',
+    claimAt: '2022-09-02T14:00:00.000Z',
+    withdrawAt: '2023-06-21T14:00:00.000Z',
+    top: 32,
+    left: 840,
+    animation: '2s 2s infinite alternate ease-in-out float'
+  },
+  {
+    id: GatewayPoolsEnum['XDEFI'],
+    rewards: 0,
+    rewardsInUst: 0,
+    rewardDenominator: 'XDEFI',
+    totalDepositAmount: 500.980584,
+    totalWithdrawnAmount: 0,
+    totalClaimedAmount: 0,
+    totalClaimedAmountInUst: 0,
+    claimedAmountToUstMultiplier: 1,
+    startedAt: "2021-12-22T14:00:00.000Z",
+    claimAt: "2022-04-22T14:00:00.000Z",
+    withdrawAt: "2022-12-22T14:00:00.000Z",
+    top: 215,
+    left: -409,
+    animation: '2s 2s infinite alternate ease-in-out float'
+  }
+];
